requirements.txt
# LLM Fundamentals - Dependencies

google-genai
python-dotenv
numpy
sentence-transformers
torch
PyPDF2
fastapi
python-multipart
uvicorn
streamlit

